Move to tests/benchmarks/
